+++ e2e/test-env/src/shared/components/Editor/Editor.jsx
@@ -2,16 +2,29 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { convertFromRaw, convertToRaw, EditorState } from '@wix/draft-js';
 import deepFreeze from 'deep-freeze';
-import { RichContentEditor } from 'wix-rich-content-editor';
+import { RichContentEditor, RichContentEditorModal } from 'wix-rich-content-editor';
 import 'wix-rich-content-common/dist/styles.min.css';
 import 'wix-rich-content-editor/dist/styles.min.css';
 import theme from '../../theme';
 import * as Plugins from './editorPlugins';
+import ReactModal from 'react-modal';
+import ModalsMap from './ModalsMap';
 
+const modalStyleDefaults = {
+  content: {
+    top: '50%',
+    left: '50%',
+    right: 'auto',
+    bottom: 'auto',
+    marginRight: '-50%',
+    transform: 'translate(-50%, -50%)',
+  },
+};
 class Editor extends Component {
   static propTypes = {
     initialState: PropTypes.object,
     isMobile: PropTypes.bool,
+    locale: PropTypes.string,
   };
 
   state = {
@@ -66,6 +79,11 @@ class Editor extends Component {
     },
   };
   render() {
+    const modalStyles = {
+      content: Object.assign({}, (this.state.modalStyles || modalStyleDefaults).content),
+      overlay: Object.assign({}, (this.state.modalStyles || modalStyleDefaults).overlay),
+    };
+    const { onRequestClose } = this.state.modalProps || {};
     return (
       <>
         Editor
@@ -78,7 +96,21 @@ class Editor extends Component {
           config={Plugins.config}
           isMobile={this.props.isMobile}
           helpers={this.helpers}
+          locale={this.props.locale}
         />
+        <ReactModal
+          isOpen={this.state.showModal}
+          contentLabel="External Modal Example"
+          style={modalStyles}
+          role="dialog"
+          onRequestClose={onRequestClose || this.helpers.closeModal}
+        >
+          <RichContentEditorModal
+            modalsMap={ModalsMap}
+            locale={this.props.locale}
+            {...this.state.modalProps}
+          />
+        </ReactModal>
       </>
     );
   }